Bugfix/ax 157 wsgi wont start until first request (#67)
* fixed bug AX-157

no tests yet

* small fix

* autopep8

* added verbose

* god-damnit!

* changed lib version

* formatted

* read-timeout is now bigger

* finally made a hack.

* formatted

* seperated axonius-libs to axonius-base-image

* more

* fix

* Added status for tests

* Added lock to serverutils and made prod configurations

* seperated

* more fixes

* fixed gui not compiling problem

* removed mapping of docker-compose for all containers

* pulling axonius-base-image from docker hub

* formatted

* make more verbose tests that fail on first test failure +x +l

* fix volatile config getting

* make everything persistent

* print logs

* some tests

* more

* skip core restart

* changed to 600, removed -x,

* now really removed -x

* Ofir fixes

* Fix Avidor's bug

* Changed to lan

* add stderr to subprocess popen

* Fixed tests

* Skipping DB check

* Formatting

* Fixed the skip issue

+++ testing/services/compose_service.py
@@ -1,14 +1,17 @@
 import subprocess
 import os
 from abc import abstractmethod
+import datetime
 
 import services.axon_service
 
 
 class ComposeService(services.axon_service.AxonService):
-    def __init__(self, compose_file_path):
+    def __init__(self, compose_file_path, container_name):
         super().__init__()
         self._compose_file_path = compose_file_path
+        self.container_name = container_name
+        self.start()
 
     def start(self):
         self.stop()
@@ -16,15 +19,50 @@ class ComposeService(services.axon_service.AxonService):
                               cwd=os.path.dirname(self._compose_file_path))
 
     def stop(self, should_delete=True):
-        subprocess.call(['docker-compose', 'stop'],
-                        cwd=os.path.dirname(self._compose_file_path))
-        if should_delete:
-            subprocess.call(['docker-compose', 'rm', "-f"],
-                            cwd=os.path.dirname(self._compose_file_path))
+        # No need to stop if its not there.
+        p = subprocess.Popen(
+            ['docker', 'ps', '--filter', 'name={0}'.format(self.container_name)], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        (out, err) = p.communicate()
+        if self.container_name in out.decode("utf-8"):
+            p = subprocess.Popen(['docker', 'logs', self.container_name],
+                                 stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+            (out, err) = p.communicate()
+            with open(os.path.join("..", "logs", "{0}_docker_logs.txt".format(self.container_name)), "a") as f:
+                f.write("====== START docker logs {0} ===========".format(self.container_name))
+                f.write("out:")
+                f.write(out.decode("utf-8"))
+                f.write("err:")
+                f.write(err.decode("utf-8"))
+                f.write("====== END   docker logs {0} ===========".format(self.container_name))
+
+            p = subprocess.call(['docker-compose', 'stop'],
+                                cwd=os.path.dirname(self._compose_file_path))
+
+            if should_delete:
+                p = subprocess.call(['docker-compose', 'rm', "-f"],
+                                    cwd=os.path.dirname(self._compose_file_path))
 
     def start_and_wait(self):
+        """
+        Take notice that the constructor already calls 'start' method. So use this function only
+        after manual stop
+        :return:
+        """
         self.start()
         self.wait_for_service()
+
+    def get_file_contents_from_container(self, file_path):
+        """
+        Gets the contents of an internal file.
+        :param file_path: the absolute path inside the container.
+        :return: the contents
+        """
+
+        p = subprocess.Popen(['docker', 'exec', self.container_name, 'cat', file_path],
+                             stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        (out, err) = p.communicate()
+
+        return (out, err, p.returncode)
 
     @abstractmethod
     def is_up(self):
